api/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
pgvector==0.2.5
minio==7.2.13
pika==1.3.2
sentence-transformers==5.3.0